// src/components/Contributors/types.ts
export interface Contributor {
  login: string;
  id: number;
  node_id: string;
  avatar_url: string;
  gravatar_id: string;
  url: string;
  html_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  starred_url: string;
  subscriptions_url: string;
  organizations_url: string;
  repos_url: string;
  events_url: string;
  received_events_url: string;
  type: string;
  site_admin: boolean;
  contributions: number;
}

export interface ContributorsProps {
  owner: string;
  repo: string | string[];
}

export interface ContributorsState {
  contributors: Contributor[];
}

export interface ProfileComponentProps {
  avatar_url: string;
  name: string;
  url: string;
  contributions: number;
}